requirements.txt
torch>=1.5.0,!=1.8
transformers>=4.0.0
matplotlib>=2.2.3
tensorboard
datasets
hydra-core
sklearn
scipy
setuptools==59.5.0
